go/ssa: fix MethodValue race
Program.MethodValue returns the wrapper function for a given
function symbol, creating them on demand and caching them.
However, methods are inserted into the cache before building
is complete, so if MethodValue is called concurrently, thread
A may construct a function, insert it in the cache, then
build it, while another thread B retrieves it from the cache
and returns without waiting, causing it to observe the function
in the process of being built.

The fix is for threads making a cache hit to wait for an
event that indicates completion of building. The event
is represented as a field, 'built chan unit'.
(The extra synchroniziation makes no apparent difference
to TestStdlib's SSA create and build metrics.)

Also:
- Add a test that reliably reproduces the data race.
- Un-export the methods created.{add,at,len}.
- rename Function.done to 'finish', and simplify.
- define type unit = struct{} and clean up.
- consolidate clean-up of transient builder state in finishBody.

Fixes golang/go#67079

Change-Id: I621150b962dc0430d78d2895d81aa7015a53e2cb